get ci pipelines functional, and switch to ruff linter/formatter

+++ enforcer/enforcer.py
@@ -1,6 +1,8 @@
 """discord red-bot enforcer"""
+
 import asyncio
-from typing import Union
+import logging
+from typing import ClassVar, Optional, Union, cast
 
 import discord
 from redbot.core import Config, checks, commands
@@ -18,11 +20,13 @@ KEY_MINGUILDAGE = "minimumguildage"
 
 CUSTOM_CONTROLS = {"⬅️": prev_page, "⏹️": close_menu, "➡️": next_page}
 
+log = logging.getLogger("red.rhomelab.enforcer")
+
 
 class EnforcerCog(commands.Cog):
     """Enforcer Cog"""
 
-    ATTRIBUTES = {
+    ATTRIBUTES: ClassVar = {
         KEY_ENABLED: {"type": "bool"},
         KEY_MINCHARS: {"type": "number"},
         KEY_MAXCHARS: {"type": "number"},
@@ -45,8 +49,16 @@ class EnforcerCog(commands.Cog):
 
         self.config.register_guild(**default_guild_settings, force_registration=True)
 
+    def _is_valid_channel(self, channel: discord.guild.GuildChannel | None):
+        if channel is not None and not isinstance(channel, (discord.ForumChannel, discord.CategoryChannel)):
+            return channel
+        return False
+
     @commands.Cog.listener()
     async def on_message(self, message: discord.Message):
+        if not isinstance(message.guild, discord.Guild):
+            # The user has DM'd us. Ignore.
+            return
         if not self.is_valid_message(message):
             return
 
@@ -64,6 +76,15 @@ class EnforcerCog(commands.Cog):
 
     @commands.Cog.listener()
     async def on_msg_enforce(self, message: discord.Message, reason: str):
+        if (
+            not isinstance(message.guild, discord.Guild)
+            or isinstance(message.channel, discord.GroupChannel)
+            or isinstance(message.channel, discord.DMChannel)
+            or isinstance(message.channel, discord.PartialMessageable)
+        ):
+            # The user has DM'd us. Ignore.
+            return
+
         await message.delete()
 
         author = message.author
@@ -76,33 +97,45 @@ class EnforcerCog(commands.Cog):
         log_id = await self.config.guild(message.guild).logchannel()
         if log_id:
             log_channel = message.guild.get_channel(log_id)
-            if log_channel:
+            if channel := self._is_valid_channel(log_channel):
                 try:
-                    await log_channel.send(embed=data)
+                    await channel.send(embed=data)
                 except discord.Forbidden:
-                    await log_channel.send(f"**Message Enforced** - {author.id} - {author} - Reason: {reason}")
+                    await channel.send(f"**Message Enforced** - {author.id} - {author} - Reason: {reason}")
+            else:
+                log.warning(
+                    f"Could not find log channel for guild {message.guild.id}, message was: **Message Enforced** "
+                    f"- {author.id} - {author} - Reason: {reason}"
+                )
 
         if not author.dm_channel:
             await author.create_dm()
-
+            dm_channel = cast(discord.DMChannel, author.dm_channel)
+        else:
+            dm_channel = author.dm_channel
         try:
-            await author.dm_channel.send(embed=data)
+            await dm_channel.send(embed=data)
         except discord.Forbidden:
             # User does not allow DMs
             inform_id = await self.config.guild(message.guild).userchannel()
             if inform_id:
                 inform_channel = message.guild.get_channel(inform_id)
-                if inform_channel:
-                    await inform_channel.send(content=author.mention, embed=data)
+                if channel := self._is_valid_channel(inform_channel):
+                    await channel.send(content=author.mention, embed=data)
+                else:
+                    log.warning(
+                        f"Could not find inform channel for guild {message.guild.id}, message was: **Message Enforced** "
+                        f"- {author.id} - {author} - Reason: {reason}"
+                    )
 
-    @commands.group(name="enforcer")
+    @commands.group(name="enforcer")  # type: ignore
     @commands.guild_only()
     @checks.admin()
     async def _enforcer(self, ctx: commands.Context):
         pass
 
     @_enforcer.command("logchannel")
-    async def enforcer_logchannel(self, ctx: commands.Context, channel: discord.TextChannel):
+    async def enforcer_logchannel(self, ctx: commands.GuildContext, channel: discord.TextChannel):
         """Sets the channel to post the enforcer logs.
 
         Example:
@@ -113,7 +146,7 @@ class EnforcerCog(commands.Cog):
         await ctx.send(f"Enforcer log message channel set to `{channel.name}`")
 
     @_enforcer.command("userchannel")
-    async def enforcer_userchannel(self, ctx: commands.Context, channel: discord.TextChannel):
+    async def enforcer_userchannel(self, ctx: commands.GuildContext, channel: discord.TextChannel):
         """Sets the channel to inform the user of deletion reason, if DMs are unavailable.
 
         Example:
@@ -130,7 +163,7 @@ class EnforcerCog(commands.Cog):
         channel: discord.TextChannel,
         attribute: str,
         *,
-        value: str = None,
+        value: Optional[str] = None,
     ):
         """Allows configuration of a channel
 
@@ -163,16 +196,16 @@ class EnforcerCog(commands.Cog):
 
         # Validate the input from the user
         try:
-            value = await self._validate_attribute_value(attribute, value)
+            validated_value = await self._validate_attribute_value(attribute, value)
         except ValueError:
             await ctx.send("The given value is invalid for that attribute.")
             return
 
-        await self._set_attribute(channel, attribute, value)
+        await self._set_attribute(channel, attribute, validated_value)
         await ctx.send(f"Channel has now configured the {attribute} attribute.")
 
     @_enforcer.command("status")
-    async def enforcer_status(self, ctx: commands.Context):
+    async def enforcer_status(self, ctx: commands.GuildContext):
         """Prints the status of the enforcement cog
 
         Example:
@@ -182,10 +215,14 @@ class EnforcerCog(commands.Cog):
         async with self.config.guild(ctx.guild).channels() as channels:
             for channel_obj in channels:
                 channel = ctx.guild.get_channel(channel_obj["id"])
-
                 conf_str = "\n".join(f"{key} - {channel_obj[key]}" for key in self.ATTRIBUTES if key in channel_obj)
-
-                messages.append(f"📝{channel.mention} - Configuration\n{conf_str}")
+                if channel:
+                    messages.append(f"📝{channel.mention} - Configuration\n{conf_str}")
+                else:
+                    messages.append(
+                        f"📝Channel ID {channel_obj['id']} no longer exists but has config, remove it... "
+                        f"- Configuration\n{conf_str}"
+                    )
 
         # Pagify implementation
         # https://github.com/Cog-Creators/Red-DiscordBot/blob/9698baf6e74f6b34f946189f05e2559a60e83706/redbot/core/utils/chat_formatting.py#L208
@@ -208,7 +245,7 @@ class EnforcerCog(commands.Cog):
                 timeout=30.0,
             )
         else:
-            ctx.send("No configurations found")
+            await ctx.send("No configurations found")
 
     async def _validate_attribute_value(self, attribute: str, value: str) -> Union[str, int, bool]:
         attribute_type = self.ATTRIBUTES[attribute]["type"]
@@ -223,9 +260,8 @@ class EnforcerCog(commands.Cog):
         if attribute_type == "number":
             if not value.isdigit():
                 raise ValueError()
-            value = int(value)
-
-            return value
+            return int(value)
+        return value
 
     async def _reset_attribute(self, channel: discord.TextChannel, attribute: str):
         async with self.config.guild(channel.guild).channels() as channels:
@@ -264,45 +300,48 @@ class EnforcerCog(commands.Cog):
     async def check_enforcer_rules(self, channel: dict, message: discord.Message) -> Union[bool, str]:
         """Check message against channel enforcer rules"""
         author = message.author
+        enforcer_error = ""
 
         if not channel.get(KEY_ENABLED):
             # Enforcing not enabled here
             return False
 
-        if KEY_MINDISCORDAGE in channel and author.created_at:
+        elif KEY_MINDISCORDAGE in channel and author.created_at:
             delta = discord.utils.utcnow() - author.created_at
             if delta.total_seconds() < channel[KEY_MINDISCORDAGE]:
                 # They breached minimum discord age
-                return "User account not old enough"
+                enforcer_error = "User account not old enough"
 
-        if KEY_MINGUILDAGE in channel and author.joined_at:
+        elif KEY_MINGUILDAGE in channel and isinstance(author, discord.Member) and author.joined_at:
             delta = discord.utils.utcnow() - author.joined_at
             if delta.total_seconds() < channel[KEY_MINGUILDAGE]:
                 # They breached minimum guild age
-                return "User not in server long enough"
+                enforcer_error = "User not in server long enough"
 
-        if channel.get(KEY_NOTEXT) and not message.content:
+        elif channel.get(KEY_NOTEXT) and not message.content:
             # They breached notext attribute
-            return "Message had no text"
+            enforcer_error = "Message had no text"
 
-        if (KEY_MINCHARS in channel) and (len(message.content) < channel[KEY_MINCHARS]):
+        elif (KEY_MINCHARS in channel) and (len(message.content) < channel[KEY_MINCHARS]):
             # They breached minchars attribute
-            return "Not enough characters"
+            enforcer_error = "Not enough characters"
 
-        if (KEY_MAXCHARS in channel) and (len(message.content) > channel[KEY_MAXCHARS]):
+        elif (KEY_MAXCHARS in channel) and (len(message.content) > channel[KEY_MAXCHARS]):
             # They breached maxchars attribute
-            return "Too many characters"
+            enforcer_error = "Too many characters"
 
-        if channel.get(KEY_NOMEDIA) or channel.get(KEY_REQUIREMEDIA):
+        elif channel.get(KEY_NOMEDIA) or channel.get(KEY_REQUIREMEDIA):
             # Check the embeds
             embeds = await self.check_embeds(message)
             if channel.get(KEY_NOMEDIA) and (embeds or message.attachments):
                 # They breached nomedia attribute
-                return "No media allowed"
+                enforcer_error = "No media allowed"
             if channel.get(KEY_REQUIREMEDIA) and not (embeds or message.attachments):
                 # They breached requiremedia attribute
-                return "Requires media attached"
+                enforcer_error = "Requires media attached"
 
+        if enforcer_error:
+            return enforcer_error
         return False
 
     async def check_embeds(self, message: discord.Message) -> bool:
